hand gesture test
test1- The swipe functionality isnt entirely correct. I think it has to do with the fact that this is ran each frame and it runs on around 30fps i believe. This meanas that unless i move my hand really fast it wont detect it as a swipe. I need to find a general sweet spot and alter the way i detect it.