NEED Camera#2 motor file
Added more stuff/logic about how everything will communicate. Still need the function for how the gesture camera(camera#2) will send its signals to its motor. *Note this motor is attached and controlled by the pi so it does not need to directly communicate with the microcontroller. However, the camera#2 needs to send back to the pi its data since it will either interrupt or keep going and it needs to happen constantly and minimal errors since it will be flying while it happens.